pdf_classifier/requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
PyMuPDF>=1.18.0
imagehash>=4.2.0

